test: adapt Csc CPU bridge coverage to latest optimizer
